backend-cli/avangcli/core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "Core Module"
authors = ["Pedro Infante <infante@avangenio.com>"]
packages = [{ include = "core", format = "sdist" }]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.116.1"
passlib = "1.7.4"
pydantic = "2.11.7"
SQLAlchemy = "2.0.41"
psycopg2-binary = "2.9.10"
uvicorn = "0.35.0"
nest-asyncio = "1.6.0"
sqlalchemy-filterset = "2.0.0"
python-decouple = ">=3.8"
inject = ">=5.1.0"
simple-settings = "1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
mypy = "^1.7.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"